load imaginary from index
